[MOVE/#107] 계좌, 주소뷰 패키징 수정
