+++ presentation/src/main/java/co/orange/presentation/delivery/DeliveryActivity.kt
@@ -11,7 +11,8 @@ import co.orange.core.extension.setOnSingleClickListener
 import co.orange.core.extension.stringOf
 import co.orange.core.extension.toast
 import co.orange.core.state.UiState
-import co.orange.presentation.delivery.AddressActivity.Companion.DEFAULT_ID
+import co.orange.presentation.delivery.address.AddressActivity
+import co.orange.presentation.delivery.address.AddressActivity.Companion.DEFAULT_ID
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.launchIn
