+++ presentation/src/main/java/co/orange/presentation/delivery/DeliveryViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import co.orange.core.state.UiState
 import co.orange.domain.entity.response.AddressModel
 import co.orange.domain.repository.SettingRepository
-import co.orange.presentation.delivery.AddressActivity.Companion.DEFAULT_ID
+import co.orange.presentation.delivery.address.AddressActivity.Companion.DEFAULT_ID
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
