+++ presentation/src/main/java/co/orange/presentation/delivery/address/AddressViewModel.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.delivery
+package co.orange.presentation.delivery.address
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
