+++ presentation/src/main/java/co/orange/presentation/delivery/address/AddressWebBridge.kt
@@ -1,4 +1,4 @@
-package co.orange.presentation.delivery
+package co.orange.presentation.delivery.address
 
 import android.app.Activity
 import android.app.Activity.RESULT_OK
