+++ presentation/src/main/res/layout/activity_address.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="co.orange.presentation.delivery.AddressViewModel" />
+            type="co.orange.presentation.delivery.address.AddressViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
