+++ presentation/src/main/res/layout/activity_bank_add.xml
@@ -9,7 +9,7 @@
 
         <variable
             name="vm"
-            type="co.orange.presentation.bank.AddBankViewModel" />
+            type="co.orange.presentation.bank.add.BankAddViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
